langgraph-mcp-baseline/pyproject.toml
[project]
name = "langgraph-mcp-baseline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "langchain-community>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain[anthropic]>=0.3.27",
    "langgraph>=0.6.6",
    "langgraph-openai-serve>=0.0.7",
    "langsmith>=0.4.14",
]
